move bulk of logic to index.ts and lint
