+++ src/launcher/plugin.ts
@@ -6,26 +6,6 @@ import {
 } from 'phosphor/lib/algorithm/json';
 
 import {
-  DisposableDelegate, IDisposable
-} from 'phosphor/lib/core/disposable';
-
-import {
-  h, VNode
-} from 'phosphor/lib/ui/vdom';
-
-import {
-  Widget
-} from 'phosphor/lib/ui/widget';
-
-import {
-  ObservableList
-} from '../common/observablelist';
-
-import {
-  VDomModel, VDomWidget
-} from '../common/vdom';
-
-import {
   JupyterLab, JupyterLabPlugin
 } from '../application';
 
@@ -42,54 +22,8 @@ import {
 } from '../services';
 
 import {
-  ILauncher
+  ILauncher, LauncherModel, LauncherWidget
 } from './';
-
-
-/**
- * The class name added to LauncherWidget instances.
- */
-const LAUNCHER_CLASS = 'jp-LauncherWidget';
-
-/**
- * The class name added to LauncherWidget image nodes.
- */
-const IMAGE_CLASS = 'jp-LauncherWidget-image';
-
-/**
- * The class name added to LauncherWidget text nodes.
- */
-const TEXT_CLASS = 'jp-LauncherWidget-text';
-
-/**
- * The class name added to LauncherWidget column nodes.
- */
-const COLUMN_CLASS = 'jp-LauncherWidget-column';
-
-/**
- * The class name added to LauncherWidget folder node.
- */
-const FOLDER_CLASS = 'jp-LauncherWidget-folder';
-
-/**
- * The class name added to LauncherWidget path nodes.
- */
-const PATH_CLASS = 'jp-LauncherWidget-path';
-
-/**
- * The class name added to LauncherWidget current working directory node.
- */
-const CWD_CLASS = 'jp-LauncherWidget-cwd';
-
-/**
- * The class name added to LauncherWidget body nodes.
- */
-const BODY_CLASS = 'jp-LauncherWidget-body';
-
-/**
- * The class name added to LauncherWidget dialog node.
- */
-const DIALOG_CLASS = 'jp-LauncherWidget-dialog';
 
 
 /**
@@ -111,14 +45,14 @@ const launcherExtension: JupyterLabPlugin<ILauncher> = {
 function activateLauncher(app: JupyterLab, services: IServiceManager, pathTracker: IPathTracker, palette: ICommandPalette): ILauncher {
   let launcherModel = new LauncherModel();
 
-  launcherModel.setDir(pathTracker.path);
-  launcherModel.setApp(app);
+  launcherModel.setDir(pathTracker.path); // change to = 
+  launcherModel.setApp(app); // change to = 
 
   pathTracker.pathChanged.connect(() => {
-    launcherModel.setDir(pathTracker.path);
+    launcherModel.setDir(pathTracker.path); // change to = 
   });
 
-  let launcherWidget = new LauncherWidget()
+  let launcherWidget = new LauncherWidget();
 
   launcherWidget.model = launcherModel;
   launcherWidget.id = 'landing-jupyterlab-widget';
@@ -137,7 +71,7 @@ function activateLauncher(app: JupyterLab, services: IServiceManager, pathTracke
     `console:create-${services.kernelspecs.default}`,
     'terminal:create-new',
     'file-operations:new-text-file',
-  ]
+  ];
 
   app.commands.addCommand('jupyterlab-launcher:add-item', {
     label: 'Add Launcher Item',
@@ -150,7 +84,7 @@ function activateLauncher(app: JupyterLab, services: IServiceManager, pathTracke
   for (let i in names) {
     // Note: we do not retain a handle on the items added by default, which
     // means we have to way of removing them after the fact.
-    app.commands.execute('jupyterlab-launcher:add-item', {name: names[i], action: actions[i]});
+    launcherModel.add(names[i], actions[i]);
   }
 
 
@@ -173,126 +107,3 @@ function activateLauncher(app: JupyterLab, services: IServiceManager, pathTracke
   return launcherModel;
 }
 
-
-/**
- * Simple encapsulation of name and callback of launcher entries.
- */
-class LauncherItem {
-
-  readonly name: string;
- 
-  readonly clickCallback: () => void;
-  
-  readonly imgName: string;
- 
-  constructor(name: string, clickCallback: () => void, imgName?: string) {
-    this.name = name;
-    this.clickCallback = clickCallback;
-    if (imgName) {
-      this.imgName = imgName;
-    } else {
-      this.imgName = 'jp-Image' + name.replace(/\ /g, '');
-    }
-  }
-}
-
-
-/**
- * LauncherModel keeps track of the path to working directory and has a list of
- * LauncherItems.
- */
-class LauncherModel extends VDomModel implements ILauncher {
-  items: LauncherItem[] = [];
-  path: string = 'home';
-  app: JupyterLab;
-  /**
-   * Convenience method to add a launcher with a given name and callback.
-   *
-   * This keeps plugins who wish to register themselves in the Launcher from
-   * having to import both LauncherItems and LauncherItem.
-   */
-  add(name: string, action: string, args?: JSONObject, imgName?: string) : IDisposable {
-    let clickCallback = () => { this.app.commands.execute( action, args) };
-    return this.addItem(new LauncherItem(name, clickCallback, imgName));
-  }
-
-  /**
-   * Add a new launcher and trigger re-render event for parent widget.
-   *
-   * @returns A Disposable which can be called to remove this new item from
-   * the Launcher and trigger another re-render event.
-   */
-  addItem(item: LauncherItem) : IDisposable {
-    this.items.push(item);
-    this.stateChanged.emit(void 0);
-
-    return new DisposableDelegate(() => {
-      // Remove the item form the list of items.
-      var index = this.items.indexOf(item, 0);
-      if (index > -1) {
-          this.items.splice(index, 1);
-          this.stateChanged.emit(void 0);
-      }
-    });
-  }
-
-  /**
-   * Set the path to the current working directory.
-   */
-  setDir(path: string) : void {
-    this.path = path;
-    this.stateChanged.emit(void 0);
-  }
-
-  /**
-   * Set the JupyterLab application this launcher will use when executing
-   * commands.
-   */
-  setApp(app: JupyterLab) : void {
-    this.app = app;
-  }
-}
-
-
-/**
- * A virtual-DOM-based widget for the Launcher.
- */
-class LauncherWidget extends VDomWidget<LauncherModel> {
-
-  /**
-   * Construct a new launcher widget.
-   */
-  constructor() {
-    super();
-    this.addClass(LAUNCHER_CLASS);
-  }
-
-  /**
-   * Render the launcher to virtual DOM nodes.
-   */
-  protected render(): VNode | VNode[] {
-    let children : VNode[] = [];
-
-    for (let item of this.model.items) {
-      let img = h.span({className: item.imgName + ' ' + IMAGE_CLASS});
-      let text = h.span({className:  TEXT_CLASS }, item.name);
-
-      let column = h.div({
-        className: COLUMN_CLASS,
-        'onclick': item.clickCallback
-      }, [img, text])
-      children.push(column);
-    }
-
-    let folderImage = h.span({ className: FOLDER_CLASS });
-    let p = this.model.path;
-    let pathName = p.length ? 'home > ' + p.replace(/\//g, ' > ') : 'home';
-    let path = h.span({ className: PATH_CLASS }, pathName );
-
-    let cwd = h.div({ className: CWD_CLASS }, [folderImage, path]);
-    let body = h.div({ className: BODY_CLASS  }, children);
-
-    return h.div({ className: DIALOG_CLASS}, [ cwd, body ]);
-
-  }
-}
